backend/requirements.txt
txtfastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0

# AI & ML
anthropic==0.18.1
langchain==0.1.6
langchain-community==0.0.20
openai==1.12.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1

# Image Processing
opencv-python==4.9.0.80
pillow==10.2.0
pytesseract==0.3.10

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
redis==5.0.1